requirements.txt
# MCP Server Framework
mcp[cli]>=1.14.0
fastapi
uvicorn

# Document Processing
docling
tiktoken
transformers

# AI/ML
openai

# Database
psycopg2-binary

# Utilities
requests
python-dotenv
pydantic
